parser: separate parser from ast tokens
