+++ src/parser/mod.rs
@@ -1,2 +1,3 @@
+mod ast;
 mod operator;
 pub mod parser;
