+++ src/parser/parser.rs
@@ -1,103 +1,7 @@
 use crate::lexer::lexer::Lexer;
+use crate::parser::ast::*;
 use crate::parser::operator::{is_bool_op, map_binary_operator};
 use crate::token::{Kind, Token, TokenValue};
-
-#[derive(Debug, Default, Clone, Copy, PartialEq, Eq)] // #[serde(tag = "type")]
-pub struct Node {
-    /// Start offset in source
-    pub start: usize,
-
-    /// End offset in source
-    pub end: usize,
-}
-
-impl Node {
-    pub fn new(start: usize, end: usize) -> Self {
-        Self { start, end }
-    }
-}
-
-// The following structs are used to represent the AST
-// https://docs.python.org/3/library/ast.html#abstract-grammar
-#[derive(Debug)]
-pub struct Module {
-    pub node: Node,
-    pub body: Vec<Statement>,
-}
-
-#[derive(Debug)]
-pub enum Statement {
-    AssignStatement(Assign),
-    ExpressionStatement(Expression),
-}
-
-#[derive(Debug)]
-pub struct Assign {
-    pub node: Node,
-    pub targets: Vec<Expression>,
-    pub value: Expression,
-}
-
-#[derive(Debug)]
-pub enum Expression {
-    Constant(Box<Constant>),
-    List(Box<List>),
-    Tuple(Box<Tuple>),
-    Name(Box<Name>),
-    BoolOp(Box<BoolOperation>),
-}
-
-// https://docs.python.org/3/reference/expressions.html#atom-identifiers
-#[derive(Debug)]
-pub struct Name {
-    pub node: Node,
-    pub id: String,
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct Constant {
-    pub node: Node,
-    pub value: ConstantValue,
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum ConstantValue {
-    None,
-    Bool(bool),
-    Str(String),
-    Bytes(Vec<u8>),
-    Tuple(Vec<Constant>),
-    // Numbers are string because we don't care about the value rn.
-    Int(String),
-    Float(String),
-    Complex { real: String, imaginary: String },
-}
-
-#[derive(Debug)]
-pub struct List {
-    pub node: Node,
-    pub elements: Vec<Expression>,
-}
-
-#[derive(Debug)]
-pub struct Tuple {
-    pub node: Node,
-    pub elements: Vec<Expression>,
-}
-
-// https://docs.python.org/3/library/ast.html#ast.BoolOp
-#[derive(Debug)]
-pub struct BoolOperation {
-    pub node: Node,
-    pub op: BooleanOperator,
-    pub values: Vec<Expression>,
-}
-
-#[derive(Debug)]
-pub enum BooleanOperator {
-    And,
-    Or,
-}
 
 #[derive(Debug)]
 pub struct Parser {
